+++ src/exponential.jl
@@ -10,7 +10,7 @@ end
 """
 Reverse function for `exp`.
 """
-function exp_rev(y::Interval, x::Interval)
+function exp_rev(y::IntervalType, x::IntervalType)
     y_new = y ⊓ (interval(0, Inf))
     x_new = x ⊓ log(y)
     return y_new, x_new
@@ -19,7 +19,7 @@ end
 """
 Reverse function for `exp2`.
 """
-function exp2_rev(y::Interval, x::Interval)
+function exp2_rev(y::IntervalType, x::IntervalType)
     y_new = y ⊓ (interval(0, Inf))
     x_new = x ⊓ log2(y)
 
@@ -29,7 +29,7 @@ end
 """
 Reverse function for `exp10`.
 """
-function exp10_rev(y::Interval, x::Interval)
+function exp10_rev(y::IntervalType, x::IntervalType)
     y_new = y ⊓ (interval(0, Inf))
     x_new = x ⊓ log10(y)
 
@@ -39,7 +39,7 @@ end
 """
 Reverse function for `expm1`.
 """
-function expm1_rev(y::Interval, x::Interval)
+function expm1_rev(y::IntervalType, x::IntervalType)
     y_new = y ⊓ (interval(-1, Inf))
     x_new = x ⊓ log1p(y)
 
@@ -59,7 +59,7 @@ end
 """
 Reverse function for `log`: ``y = \\log(x)``
 """
-function log_rev(y::Interval, x::Interval)
+function log_rev(y::IntervalType, x::IntervalType)
     x_new = x ⊓ exp(y)
 
     return y, x_new
@@ -68,7 +68,7 @@ end
 """
 Reverse function for `log2`: ``y = \\log2(x)``
 """
-function log2_rev(y::Interval, x::Interval)
+function log2_rev(y::IntervalType, x::IntervalType)
     x_new = x ⊓ exp2(y)
 
     return y, x_new
@@ -78,7 +78,7 @@ end
 """
 Reverse function for `log10`: ``y = \\log10(x)``
 """
-function log10_rev(y::Interval, x::Interval)
+function log10_rev(y::IntervalType, x::IntervalType)
     x_new = x ⊓ exp10(y)
 
     return y, x_new
@@ -87,7 +87,7 @@ end
 """
 Reverse function for `log1p`: ``y = \\log1p(x)``
 """
-function log1p_rev(y::Interval, x::Interval)
+function log1p_rev(y::IntervalType, x::IntervalType)
     x_new = x ⊓ expm1(y)
 
     return y, x_new