+++ src/transformations.jl
@@ -1,27 +1,28 @@
 
 """
-    integer_contractor(x::Interval)
+    integer_contractor(x::IntervalType)
 
 Return the integers enclosed in the interval `x`.
 """
 
-function integer_contractor(x::Interval)
-    a = floor(inf(x))+1
+function integer_contractor(x::IntervalType)
+    a = floor(inf(x)) + 1
     b = floor(sup(x))
 
     a > b && return emptyinterval(x)
 
-    return interval(a, b)
+    return _build_interval(x, a, b)
 end
 
 
 ### Transformations on IntervalBoxes
 
-"""Reflect in mirror at position x_mirror"""
+"""Reflect in mirror at position x_mirror
+x_mirror is a function that returns an interval giving the position of the mirror."""
 function reflect_x(x_mirror)
     X -> begin
         x, y = X
-        x = 2*x_mirror - x
+        x = (ExactReal(2) * x_mirror(x)) - x
 
         return IntervalBox(x, y)
     end
@@ -62,23 +63,23 @@ symmetrise(C, op) = op ∘ C ∘ op
 
 
 
-"Periodize the contractor C"
+"Periodize the contractor C. period is a function that returns an interval giving the period"
 function periodise(C, period)
 
     X -> begin
         x, y = X
 
-        x2 = entireinterval()
+        x2 = entireinterval(x)
         x2, y = C(IntervalBox(x2, y))
 
-        isempty(IntervalBox(x2, y)) && return(IntervalBox(emptyinterval(), emptyinterval()))
+        isempty(IntervalBox(x2, y)) && return(IntervalBox(emptyinterval(x), emptyinterval(x)))
 
         # periods where the periodization of x intersects with x2:
-        periods = integer_contractor((x - x2) / period)
+        periods = integer_contractor((x - x2) / period(x))
 
-        isempty_interval(periods) && return(IntervalBox(emptyinterval(), emptyinterval()))
+        isempty_interval(periods) && return(IntervalBox(emptyinterval(x), emptyinterval(x)))
 
-        x3 = x2 + periods*period
+        x3 = x2 + periods * period(x)
         x = x ⊓ x3
 
         return IntervalBox(x, y)